switching to schema swap
